+++ i6_models/assemblies/conformer/conformer_v2.py
@@ -5,10 +5,10 @@ __all__ = ["ConformerBlockV2Config", "ConformerEncoderV2Config", "ConformerBlock
 import torch
 from torch import nn
 from dataclasses import dataclass
-from typing import Tuple, List, Dict
+from typing import Tuple, List, Optional
 
 from i6_models.config import ModelConfiguration, ModuleFactoryV1
-from i6_models.parts.conformer import (
+from i6_models.assemblies.conformer import (
     ConformerConvolutionV1,
     ConformerConvolutionV1Config,
     ConformerMHSAV1,
@@ -25,14 +25,23 @@ class ConformerBlockV2Config(ModelConfiguration):
         ff_cfg: Configuration for ConformerPositionwiseFeedForwardV1
         mhsa_cfg: Configuration for ConformerMHSAV1
         conv_cfg: Configuration for ConformerConvolutionV1
-        swap_mhsa_conv: swap the execution order of MHSA and Conv module
+        modules: List of modules to use for ConformerBlockV2,
+                       "ff" for feed forward module, "mhsa" for multi-head self attention module, "conv" for conv module
+        scales: List of scales to apply to the module outputs before the residual connection
     """
 
     # nested configurations
     ff_cfg: ConformerPositionwiseFeedForwardV1Config
     mhsa_cfg: ConformerMHSAV1Config
     conv_cfg: ConformerConvolutionV1Config
-    swap_mhsa_conv: bool = False
+    modules: List[str] = ["ff", "mhsa", "conv", "ff"]
+    scales: List[float] = [0.5, 1.0, 1.0, 0.5]
+
+    def __post__init__(self):
+        super().__post_init__()
+        assert len(self.modules) == len(self.scales), "modules and scales must have same length"
+        for module_name in self.modules:
+            assert module_name in ["ff", "mhsa", "conv"], "module not supported"
 
 
 class ConformerBlockV2(nn.Module):
@@ -45,27 +54,34 @@ class ConformerBlockV2(nn.Module):
         :param cfg: conformer block configuration with subunits for the different conformer parts
         """
         super().__init__()
-        self.ff1 = ConformerPositionwiseFeedForwardV1(cfg=cfg.ff_cfg)
-        self.mhsa = ConformerMHSAV1(cfg=cfg.mhsa_cfg)
-        self.conv = ConformerConvolutionV1(model_cfg=cfg.conv_cfg)
-        self.ff2 = ConformerPositionwiseFeedForwardV1(cfg=cfg.ff_cfg)
+
+        modules = []
+        for module_name in cfg.modules:
+            if module_name == "ff":
+                modules.append(ConformerPositionwiseFeedForwardV1(cfg=cfg.ff_cfg))
+            elif module_name == "mhsa":
+                modules.append(ConformerMHSAV1(cfg=cfg.mhsa_cfg))
+            elif module_name == "conv":
+                modules.append(ConformerConvolutionV1(cfg=cfg.conv_cfg))
+            else:
+                raise NotImplementedError
+
+        self.modules = nn.ModuleList(modules)
+        self.scales = cfg.scales
         self.final_layer_norm = torch.nn.LayerNorm(cfg.ff_cfg.input_dim)
-        self.swap_mhsa_conv = cfg.swap_mhsa_conv
 
     def forward(self, x: torch.Tensor, /, sequence_mask: torch.Tensor) -> torch.Tensor:
         """
         :param x: input tensor of shape [B, T, F]
-        :param sequence_mask: mask tensor where 1 defines positions within the sequence and 0 outside, shape: [B, T]
+        :param sequence_mask: mask tensor where 0 defines positions within the sequence and 1 outside, shape: [B, T]
         :return: torch.Tensor of shape [B, T, F]
         """
-        x = 0.5 * self.ff1(x) + x  #  [B, T, F]
-        if not self.swap_mhsa_conv:
-            x = self.mhsa(x, sequence_mask) + x  #  [B, T, F]
-            x = self.conv(x) + x  #  [B, T, F]
-        else:
-            x = self.conv(x) + x  # [B, T, F]
-            x = self.mhsa(x, sequence_mask) + x  # [B, T, F]
-        x = 0.5 * self.ff2(x) + x  #  [B, T, F]
+        for scale, module in zip(self.scales, self.modules):
+            if isinstance(module, ConformerMHSAV1):
+                x = scale * module(x, sequence_mask) + x
+            else:
+                x = scale * module(x)
+
         x = self.final_layer_norm(x)  #  [B, T, F]
         return x
 
@@ -75,13 +91,11 @@ class ConformerEncoderV2Config(ModelConfiguration):
     """
     Attributes:
         num_layers: Number of conformer layers in the conformer encoder
-        loss_layers: Dictionary defining the loss layers e.g. [4, 8, 12]
         frontend: A pair of ConformerFrontend and corresponding config
         block_cfg: Configuration for ConformerBlockV2
     """
 
     num_layers: int
-    loss_layers: List[int]
 
     # nested configurations
     frontend: ModuleFactoryV1
@@ -90,11 +104,10 @@ class ConformerEncoderV2Config(ModelConfiguration):
 
 class ConformerEncoderV2(nn.Module):
     """
-    Modification compared to ConformerEncoderV1:
-    1. enable adding auxiliary losses with different loss scales
-        (C.f. https://arxiv.org/abs/2102.03216 for more details of auxiliary loss)
-    2. enable swapping the MHSA and Conv module
-        (C.f. https://arxiv.org/abs/2011.10798 for intuition behind swapping)
+    Implementation of the convolution-augmented Transformer (short Conformer), as in the original publication.
+    The model consists of a frontend and a stack of N conformer blocks.
+    C.f. https://arxiv.org/pdf/2005.08100.pdf
+    Each conformer block is composed of fixed number of feed forward modules, multi-head self attention modules and conv modules.
     """
 
     def __init__(self, cfg: ConformerEncoderV2Config):
@@ -106,14 +119,14 @@ class ConformerEncoderV2(nn.Module):
         self.frontend = cfg.frontend()
         self.module_list = torch.nn.ModuleList([ConformerBlockV2(cfg.block_cfg) for _ in range(cfg.num_layers)])
         self.output_layers = cfg.loss_layers
-        assert cfg.num_layers in self.output_layers, "The final layer must be included in loss layers"
 
     def forward(
-        self, data_tensor: torch.Tensor, /, sequence_mask: torch.Tensor
+        self, data_tensor: torch.Tensor, /, sequence_mask: torch.Tensor, return_layers: Optional[List[int]] = None
     ) -> Tuple[List[torch.Tensor], torch.Tensor]:
         """
         :param data_tensor: input tensor of shape [B, T', F]
         :param sequence_mask: mask tensor where 1 defines positions within the sequence and 0 outside, shape: [B, T']
+        :param return_layers: list of layer indices specifying which layers to return
         :return: (outputs, out_seq_mask)
             where outputs is a list of torch.Tensor of shape [B, T, F']
             for each of the layers in output_layers,
@@ -122,12 +135,16 @@ class ConformerEncoderV2(nn.Module):
         F: input feature dim, F': internal and output feature dim
         T': data time dim, T: down-sampled time dim (internal time dim)
         """
+
+        if return_layers is None:
+            return_layers = [len(self.module_list)]
         x, sequence_mask = self.frontend(data_tensor, sequence_mask)  # [B, T, F']
 
         outputs = []
-        for i in range(len(self.module_list)):
+        assert max(return_layers) <= len(self.module_list)
+        for i in range(max(return_layers)):
             x = self.module_list[i](x, sequence_mask)  # [B, T, F']
-            if i + 1 in self.output_layers:
+            if i + 1 in return_layers:
                 outputs.append(x)
 
         return outputs, sequence_mask
